Move `DegitErrorCode` from `index.ts` type to `utils.ts`
  - This was done to prevent potential circular imports.
